// fe_inventarios_gei/src/app/models/estadoPeriodo.ts
export class EstadoPeriodo {
  oPeriodo: any = {};
  nIdEstadoPeriodo: Number;
  nIdPeriodo: Number;
  nEstadoActual: Number;
  sEstadoActual: String;
  sDescripcion: any;
  sUIDdocumento: String;
  sNombreDocumento: String;
  sDocumento: String;
  nIdUsuarioReg: Number;
  nIdSesion: Number;
  sUsuario: String;
}